refactor: rename users dir
