+++ src/auth/auth.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
-import { UserModule } from '@/users';
+import { UserModule } from '@/user';
 import { HttpBasicStrategy } from './strategies/http-basic.strategy';
 
 @Module({
